1.分离zerosx-resource资源微服务 2.规范部分代码命名 3.修复encrypt组件验证BUG（丢失自增ID等一些字段的值） 4.导出功能增加nacos配置项 5.seata分布式事务案例优化 6.多租户前后逻辑优化（针对一个账户管理一个租户和多个租户情况） 7.个人中心模块优化[头像上传、基本信息展示更多信息] 8.补全更新md文档、数据库及YAML配置脚本 9.图片上传组件的删除按钮在删除时会删除服务器oss源文件 10.修复数据字典翻译feign接口BUG

+++ zerosx-examples/zerosx-account/src/main/java/com/zerosx/account/service/impl/AccountServiceImpl.java
@@ -11,8 +11,6 @@ import io.seata.core.context.RootContext;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
-import java.math.BigDecimal;
-
 @Service
 @Slf4j
 public class AccountServiceImpl extends ServiceImpl<IAccountMapper, Account> implements IAccountService {
@@ -23,7 +21,7 @@ public class AccountServiceImpl extends ServiceImpl<IAccountMapper, Account> imp
         LambdaQueryWrapper<Account> qw = Wrappers.lambdaQuery(Account.class);
         qw.eq(Account::getUserId, accountDTO.getUserId());
         Account dbAccount = getOne(qw);
-        dbAccount.setAmount(BigDecimal.valueOf(dbAccount.getAmount()).subtract(accountDTO.getAmount()).doubleValue());
+        dbAccount.setAmount(dbAccount.getAmount().subtract(accountDTO.getAmount()));
         return updateById(dbAccount);
     }
